prod

+++ services/sesssionServices.js
@@ -48,7 +48,13 @@ export async function updateSession(sessionId, sessionData, updatedBy) {
   return await sessionCreationService.updateSession(sessionId, sessionData);
 }
 
-export async function deleteSession(sessionId) {
+export async function deleteSession(sessionId, instituteId, universityId) {
+  const isSessionAlreadyMapped = await sessionCreationService.isSessionMappedwithcourse(sessionId, instituteId, universityId);
+
+  if (isSessionAlreadyMapped.length > 0) {
+    throw new Error('Session already mapped with courses unable to delete')
+  } 
+
   return await sessionCreationService.deleteSession(sessionId);
 };
 
@@ -56,6 +62,11 @@ export async function deleteSession(sessionId) {
 
 export async function couseSessionMapping(data, createdBy, updatedBy, universityId, instituteId) {
   try {
+    const isSessionAlreadyMapped = await sessionCreationService.isSessionAlreadyMapped(data.sessionId, data.courseId, instituteId, universityId);
+    // console.log("isSessionAlreadyMapped ", isSessionAlreadyMapped)
+    if (isSessionAlreadyMapped) {
+      throw new Error('Session already mapped')
+    }
 
     if (!Array.isArray(data.courseId) || data.courseId.length === 0) {
       throw new Error("courseId must be a non-empty array");
